osd: Make casting to/from shard_id_t require explicit casts.
This also required putting in many explicit casts.  I avoided large amount of refactoring to the new shard for now,
although I expect some may follow.  The main aim here is that the new EC code is strict about using the correct shard_id_t
and that confusion between shards, raw shards, OSDs, etc.. flag up as errors.

+++ src/crimson/osd/scrub/scrub_validator.cc
@@ -306,7 +306,7 @@ object_evaluation_t evaluate_object(
 		  [](auto &cand) { return cand.has_errors(); })) {
     for (auto &eval : shards) {
       iow.shards.emplace(
-	librados::osd_shard_t{eval.source.osd, eval.source.shard},
+	librados::osd_shard_t{eval.source.osd, static_cast<int8_t>(eval.source.shard)},
 	eval.shard_info);
       iow.union_shards.errors |= eval.shard_info.errors;
     }
